--- apps/server/src/middlewares/validators/schemas/fis.ts
@@ -1,6 +1,0 @@
-import { z } from "zod";
-import { fiSchema } from "./commons";
-
-export const getFiByIdSchema = z.object({
-  id: fiSchema,
-});
